% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_offset.R
\name{adjust_offset}
\alias{adjust_offset}
\title{Adjust Offset for RPF}
\usage{
adjust_offset(offset_df = NULL, qc_df = NULL)
}
\arguments{
\item{offset_df}{A data frame containing the read lengths and offsets.}

\item{qc_df}{A data frame containing quality control data with sample, length,
relative start positions, and other relevant information.}
}
\value{
A data frame with adjusted relative start positions.
}
\description{
This function adjusts the offsets for quality control data based on the read lengths and associated offsets.
It separates the read lengths and offsets, performs necessary data type conversions, and then adjusts
the relative start positions by the absolute values of the offsets.
}
