% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuncs.R
\name{batch_adpator_remove}
\alias{batch_adpator_remove}
\title{Batch Adapter Removal using Rfastp}
\usage{
batch_adpator_remove(
  fastq_file1 = NULL,
  fastq_file2 = NULL,
  output_dir = NULL,
  output_name = NULL,
  fastp_params = list()
)
}
\arguments{
\item{fastq_file1}{A vector of file paths to the first read in each pair.}

\item{fastq_file2}{A vector of file paths to the second read in each pair.
Default is NULL, indicating single-end reads.}

\item{output_dir}{The directory to save the output files in.}

\item{output_name}{A vector of names for the output files. Must have the same
length as fastq_file1 and fastq_file2.}

\item{fastp_params}{A list of parameters to pass to Rfastp.
See https://github.com/lh3/fastp for more information.}
}
\value{
A list of QC summaries for each input file.
}
\description{
This function removes adapters from a batch of paired-end fastq files using Rfastp.
}
\examples{
\dontrun{
# Batch adapter removal for paired-end reads
batch_adapter_remove(fastq_file1 = c("read1_1.fastq", "read1_2.fastq"),
                     fastq_file2 = c("read2_1.fastq", "read2_2.fastq"),
                     output_dir = "output/",
                     output_name = c("out1.fastq", "out2.fastq"))

# Batch adapter removal for single-end reads
batch_adapter_remove(fastq_file1 = c("read1.fastq", "read2.fastq"),
                     output_dir = "output/",
                     output_name = c("out1.fastq", "out2.fastq"))
}

}
