% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuncs.R
\name{batch_bam2bigwig}
\alias{batch_bam2bigwig}
\title{Convert BAM files to bigWig format}
\usage{
batch_bam2bigwig(bam_file = NULL, bw_file = NULL, paired = FALSE)
}
\arguments{
\item{bam_file}{A character vector giving the paths to the input BAM files.}

\item{bw_file}{A character vector giving the paths to the output bigWig files.}

\item{paired}{The sequencing type of reads, defaults False(single-end).}
}
\value{
The function invisibly returns a list containing the output file paths.
}
\description{
This function takes a set of BAM files and converts them to bigWig format.
The BAM files are read in using the \code{GenomicAlignments} package, extended
to an appropriate fragment length, converted to a GRanges object, and coverage
is calculated. The coverage data is then normalized to reads per million (RPM)
and exported as a bigWig file using the \code{rtracklayer} package.
}
\details{
The function uses the "readGAlignments" function from the
\code{GenomicAlignments} package to read in each BAM file. The resulting
object is then converted to a GRanges object using the "as" function. Coverage
is obtained using the "coverage" function from the \code{IRanges} package.
The coverage data is then normalized to RPM by dividing by the total number of
reads and multiplying by one million. The normalized data is stored as a
SimpleRleList object, which is then exported to a bigWig file using the
\code{export.bw} function from the \code{rtracklayer} package.
}
\examples{
\dontrun{
# Convert a single BAM file to bigWig format
batch_bam2bigwig(bam_file = "/path/to/my.bam",
                 bw_file = "/path/to/my.bw")

# Convert multiple BAM files to bigWig format
batch_bam2bigwig(bam_file = c("/path/to/my1.bam", "/path/to/my2.bam"),
                 bw_file = c("/path/to/my1.bw", "/path/to/my2.bw"))
}

}
