% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuncs.R
\name{batch_hisat_align}
\alias{batch_hisat_align}
\title{Batch align FASTQ files using HISAT2}
\usage{
batch_hisat_align(
  index = NULL,
  fq_file1 = NULL,
  fq_file2 = NULL,
  output_dir = NULL,
  output_file = NULL,
  threads = 1,
  hisat2_params = list()
)
}
\arguments{
\item{index}{A character string specifying the path to the HISAT2 index.}

\item{fq_file1}{A character vector specifying the paths to the first input
FASTQ files.}

\item{fq_file2}{A character vector specifying the paths to the second input
FASTQ files (optional).}

\item{output_dir}{A character string specifying the path to the output directory.}

\item{output_file}{A character vector specifying the names for the output files.}

\item{threads}{An integer specifying the number of threads to use in the
alignment process (default: 1).}

\item{hisat2_params}{A list specifying additional parameters for HISAT2
(default: empty list).}
}
\description{
This function performs batch alignment of FASTQ files using HISAT2.
}
\examples{
\dontrun{
# Align a single-end FASTQ file
batch_hisat_align(index = "path/to/index",
                  fq_file1 = "path/to/fq_file.fastq.gz",
                  output_dir = "path/to/output/",
                  output_file = "aligned.sam")

# Align paired-end FASTQ files
batch_hisat_align(index = "path/to/index",
                  fq_file1 = c("path/to/fq_file1_1.fastq.gz", "path/to/fq_file2_1.fastq.gz"),
                  fq_file2 = c("path/to/fq_file1_2.fastq.gz", "path/to/fq_file2_2.fastq.gz"),
                  output_dir = "path/to/output/",
                  output_file = c("aligned_1.sam", "aligned_2.sam"))
}

}
