% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuncs.R
\name{batch_sam2bam}
\alias{batch_sam2bam}
\title{Convert SAM files to BAM format}
\usage{
batch_sam2bam(sam_file = NULL, bam_file = NULL)
}
\arguments{
\item{sam_file}{Character vector containing the path(s) to the input SAM file(s).}

\item{bam_file}{Character vector containing the path(s) to where the output
BAM file(s) will be saved.}
}
\value{
None
}
\description{
This function converts one or more SAM files to BAM format using
Rsamtools::asBam function.
}
\examples{
\dontrun{
# Convert a single SAM file to BAM format
batch_sam2bam("my_sam_file.sam", "my_bam_file")

# Convert multiple SAM files to BAM format
batch_sam2bam(c("sam_file1.sam", "sam_file2.sam"), c("bam_file1", "bam_file2"))
}

}
