% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuncs.R
\name{bowtie2_align}
\alias{bowtie2_align}
\title{Bowtie2 Alignment}
\usage{
bowtie2_align(
  index = NULL,
  fq_file1 = NULL,
  fq_file2 = NULL,
  output_file = NULL,
  threads = 1,
  bowtie2_params = NULL
)
}
\arguments{
\item{index}{Path to the Bowtie2 index files for the reference genome.}

\item{fq_file1}{Path to the input FASTQ file containing the first set of
paired-end reads or single-end reads.}

\item{fq_file2}{Path to the input FASTQ file containing the second set of
paired-end reads. Set to NULL if analyzing single-end reads.}

\item{output_file}{Path for the output SAM file. An additional text file
containing mapping information will also be generated with the same file name
but with "_mapinfo" appended to the end.}

\item{threads}{Number of CPU threads to use for alignment.}

\item{bowtie2_params}{Additional Bowtie2 parameters to be included in the command line.}
}
\value{
Prints a message indicating that the output file has been generated.
}
\description{
This function performs Bowtie2 alignment of sequencing reads to a reference
genome. It requires the path to the Bowtie2 index files, input FASTQ files,
output file name, number of threads, and Bowtie2 parameters as inputs. If the
input FASTQ files are compressed, it automatically decompresses them before
alignment. The output includes both the mapping result in SAM format and the
mapping information in a separate text file.
}
