% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_polarity_sore.R
\name{calculate_polarity}
\alias{calculate_polarity}
\alias{calculate_polarity,ribosomeObj-method}
\title{Calculate Polarity Scores}
\usage{
calculate_polarity(object, minCounts = 64, group = NULL, ...)

\S4method{calculate_polarity}{ribosomeObj}(object, minCounts = 64, group = NULL, ...)
}
\arguments{
\item{object}{ribosomeObj object.}

\item{minCounts}{Integer. Minimum number of counts required in the CDS region to include a gene. Default is 64.}

\item{group}{Character. A grouping variable for the output. Default is NULL, in which case the group is set to the current sample.}

\item{...}{Useless args.}
}
\value{
A data frame with calculated polarity scores for each gene.
}
\description{
This function calculates polarity scores for gene expression data, filtered by CDS length.
}
\details{
The function processes gene expression data by calculating a polarity score for each gene. It filters the data
by CDS length and counts, then computes the score based on the specified parameters.
}
\examples{
\dontrun{
calculatePolarity2(object = ribosomeObj)
}

}
