% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolaritySore.R
\name{calculatePolarity}
\alias{calculatePolarity}
\title{Calculate Polarity Scores}
\usage{
calculatePolarity(
  gene_anno_file = NULL,
  gtf_file = NULL,
  input_file = NULL,
  density_file = NULL,
  minCounts = 64,
  upstreamNt = 15,
  downstreamNt = 15,
  group = NULL
)
}
\arguments{
\item{gene_anno_file}{A character string specifying the path to the gene annotation
file. Default is NULL.}

\item{gtf_file}{A character string specifying the path to the GTF file. Default is
NULL.}

\item{input_file}{A character vector specifying the paths to the SAM/BAM files. Default is
NULL.}

\item{density_file}{A character vector specifying the paths to the density files.
Default is NULL.}

\item{minCounts}{A numeric value specifying the minimum counts threshold for
filtering the counts. Default is 64.}

\item{upstreamNt}{A numeric value specifying the number of nucleotides upstream
of the UTR5 region to include. Default is 15.}

\item{downstreamNt}{A numeric value specifying the number of nucleotides downstream
of the CDS region to include. Default is 15.}

\item{group}{A character vector specifying a grouping factor. Default is NULL.}
}
\value{
A data frame containing the polarity scores.
}
\description{
This function calculates polarity scores for genes based on RNA-seq data. The
function reads gene annotation, GTF file, SAM file, density file, and optionally
a grouping factor. It then iterates over the provided SAM files, calculates counts
for each gene in the CDS region, filters the counts to those above a threshold,
and calculates the polarity score (PS) for each gene.
}
