% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolaritySore.R
\name{calculatePolarity2}
\alias{calculatePolarity2}
\title{Calculate Polarity Scores}
\usage{
calculatePolarity2(
  longest_trans_file = NULL,
  normed_file = NULL,
  minCounts = 64,
  group = NULL
)
}
\arguments{
\item{longest_trans_file}{Character. Path to a file containing gene annotation data. Default is NULL.}

\item{normed_file}{Data frame. Normalized expression data which includes `sample`, `trans_id`, `trans_pos`, and `counts`.}

\item{minCounts}{Integer. Minimum number of counts required in the CDS region to include a gene. Default is 64.}

\item{group}{Character. A grouping variable for the output. Default is NULL, in which case the group is set to the current sample.}
}
\value{
A data frame with calculated polarity scores for each gene.
}
\description{
This function calculates polarity scores for gene expression data, filtered by CDS length.
}
\details{
The function processes gene expression data by calculating a polarity score for each gene. It filters the data
by CDS length and counts, then computes the score based on the specified parameters.
}
\examples{
\dontrun{
calculatePolarity2(longest_trans_file = "path/to/annotation.txt", normed_file = expression_data)
}

}
