% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuncs.R
\name{call_cmd}
\alias{call_cmd}
\title{Call Command Function}
\usage{
call_cmd(soft_path = NULL, call_func = c("system2", "system"), ...)
}
\arguments{
\item{soft_path}{A character string that specifies the path to the software or
program that will be executed.}

\item{call_func}{A character string that specifies which function will be used
to execute the command. Allowed values are "system2" or "system".}

\item{...}{Additional arguments that will be passed to the command as a single
character string.}
}
\value{
The output of the function will depend on which function was used to
execute the command. If the system function was used, the returned value will
be a character vector with the command's output. If the system2 function was
used, the returned value will depend on the function's options.
}
\description{
This function allows the user to execute a specific command using either the
system or system2 function from the \code{base} package.
}
\examples{
\dontrun{
# Call system2 function
call_cmd(soft_path = "ls", call_func = "system2", args_cmd = "-la")

# Call system function
call_cmd(soft_path = "ls", call_func = "system", "-la")
}

}
