% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ribosomeObj.R
\name{construct_ribosomeObj}
\alias{construct_ribosomeObj}
\title{Construct a ribosomeObj Object}
\usage{
construct_ribosomeObj(
  gtf_file = NULL,
  genome_file = NULL,
  extend_utr = FALSE,
  upstream_scale = NULL,
  upstream_extend = NULL,
  downstream_scale = NULL,
  downstream_extend = NULL,
  has_created_data = FALSE,
  has_created_annotation = FALSE,
  XAM_version = c("3", "4"),
  mapping_type = c("genome", "transcriptome"),
  sam_file = NULL,
  bam_file = NULL,
  out_file_dir = NULL,
  out_file_prefix = NULL,
  annotation_prefix = NULL,
  rep_name = NULL,
  group_name = NULL,
  assignType = c("end5", "end3"),
  seq_type = c("pairedEnd", "singleEnd")
)
}
\arguments{
\item{gtf_file}{A character string specifying the path to the GTF file containing gene annotations.
This file is required for preparing gene annotation data.}

\item{genome_file}{A character string specifying the path of genome sequence.}

\item{extend_utr}{A logical value indicating whether to extend the UTR regions of genes.
Defaults to \code{FALSE}.}

\item{upstream_scale}{A numeric value specifying the scaling factor for the upstream UTR extension.
Used if \code{extend_utr = TRUE}.}

\item{upstream_extend}{A numeric value specifying how much to extend the upstream UTR.
Used if \code{extend_utr = TRUE}.}

\item{downstream_scale}{A numeric value specifying the scaling factor for the downstream UTR extension.
Used if \code{extend_utr = TRUE}.}

\item{downstream_extend}{A numeric value specifying how much to extend the downstream UTR.
Used if \code{extend_utr = TRUE}.}

\item{has_created_data}{whether you have created raw data and save time to renalysis.
Defaluts FALSE.}

\item{has_created_annotation}{whether you have created lognest gene annotation and save time to renalysis.
Defaluts FALSE.}

\item{XAM_version}{The XAM version you have installed, "3"(0.3.1) or "4"(0.4.0).}

\item{mapping_type}{A character string indicating the type of mapping, either "genome" or "transcriptome".
Defaults to "genome".}

\item{sam_file}{A character string specifying the path to the SAM file containing ribosome profiling
alignment data. Either \code{sam_file} or \code{bam_file} must be provided.}

\item{bam_file}{A character string specifying the path to the BAM file containing ribosome profiling
alignment data. Either \code{bam_file} or \code{sam_file} must be provided.}

\item{out_file_dir}{A character vector specifying the paths to the output data.}

\item{out_file_prefix}{A character string specifying the prefix for the output files generated during
the QC analysis.}

\item{annotation_prefix}{A character string specifying the prefix for the longest annotation output files.
Defaluts NULL.}

\item{rep_name}{The replicates name. Default NULL.}

\item{group_name}{The group name. Default NULL.}

\item{assignType}{A character string specifying the assignment type for mapping data,
either "end5" or "end3". Defaults to "end5".}

\item{seq_type}{A character string specifying the sequencing type, either "pairedEnd" or "singleEnd".
Defaults to "singleEnd".}
}
\value{
An S4 object of class \code{ribosomeObj}, which contains the following slots:
\item{raw.counts}{A data.frame containing the raw ribosome profiling counts.}
\item{shifted.data}{An empty data.frame for future shifted data (not yet used in this function).}
\item{normalized.data}{An empty data.frame for future normalized data (not yet used in this function).}
\item{longest.annotation}{A data.frame containing the longest annotated transcripts, potentially extended UTRs.}
}
\description{
This function constructs an S4 object of class \code{ribosomeObj}, which contains ribosome profiling data
and associated gene annotations. It takes as input a gene annotation file, and optionally allows for UTR extension
and selection of mapping, sequencing, and assignment types.
}
\details{
This function performs several steps to construct a \code{ribosomeObj} object:
\itemize{
  \item It prepares gene annotation data from the provided GTF file.
  \item It optionally extends UTR regions based on the specified parameters.
  \item It performs QC analysis on the ribosome profiling data.
  \item It loads the annotation and QC data into a \code{ribosomeObj} object.
}

The function relies on several helper functions, including \code{pre_longest_trans_info}, \code{gene_anno_extend}, \code{pre_qc_data}, and \code{load_qc_data}, to perform various tasks related to annotation processing and QC analysis.
}
\examples{
\dontrun{
# Example usage of construct_ribosomeObj
ribo <- construct_ribosomeObj(
  gtf_file = "path/to/gtf_file.gtf",
  mapping_type = "genome",
  sam_file = "path/to/alignment.sam",
  out_file_prefix = "output/ribo"
)}

}
