% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuncs.R
\name{count_to_tpm}
\alias{count_to_tpm}
\title{Calculate TPM from feature counts object}
\usage{
count_to_tpm(fc_obj = NULL)
}
\arguments{
\item{fc_obj}{Feature counts object containing count data and gene annotation
which produced from subread featureCounts function.}
}
\value{
A data frame containing gene IDs, gene names, gene biotypes, and TPM values.
}
\description{
This function takes a feature counts object and calculates the transcripts per
million (TPM) for each gene.
}
\examples{

\dontrun{
# Calculate TPM using count_to_tpm()
tpm_df <- count_to_tpm(example_fc_obj)
}

}
