% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{enrichment_track_plot}
\alias{enrichment_track_plot}
\alias{enrichment_track_plot,ribosomeObj-method}
\title{Generate Ribosome Enrichment Track Plot}
\usage{
enrichment_track_plot(
  object,
  select_gene = NULL,
  select_sample = NULL,
  assay = c("normalized", "enriched"),
  geom_type = c("line", "col"),
  filter_cds_divided_by3 = TRUE,
  merge_rep = TRUE,
  collapse = FALSE,
  sample_color = NULL,
  average_smooth = TRUE,
  window_size = 7,
  mode = c("codon", "nt"),
  add_frame_anno = FALSE,
  frame_relheight = 0.1,
  seq_mark = c("both", "start", "stop"),
  cds_region_anno = NULL,
  cds_col = "grey",
  cds_anno_col = "#996600",
  cds_region_width = 5,
  cds_region_scale_x = NULL,
  geom_line_list = list(),
  geom_col_list = list(),
  facet_grid_list = list(),
  ...
)

\S4method{enrichment_track_plot}{ribosomeObj}(
  object,
  select_gene = NULL,
  select_sample = NULL,
  assay = c("normalized", "enriched"),
  geom_type = c("line", "col"),
  filter_cds_divided_by3 = TRUE,
  merge_rep = TRUE,
  collapse = FALSE,
  sample_color = NULL,
  average_smooth = TRUE,
  window_size = 7,
  mode = c("codon", "nt"),
  add_frame_anno = FALSE,
  frame_relheight = 0.1,
  seq_mark = c("both", "start", "stop"),
  cds_region_anno = NULL,
  cds_col = "grey",
  cds_anno_col = "#996600",
  cds_region_width = 5,
  cds_region_scale_x = NULL,
  geom_line_list = list(),
  geom_col_list = list(),
  facet_grid_list = list(),
  ...
)
}
\arguments{
\item{object}{An object of class \code{ribosomeObj}, the ribosome profiling object
that contains gene annotations, normalized data, and enriched data.}

\item{select_gene}{A character vector specifying the gene name(s) to plot.}

\item{select_sample}{A character vector specifying the sample name(s) to plot.
Default is \code{NULL}, which plots all available samples.}

\item{assay}{A character string specifying whether to plot ribosome footprint density
or enrichment data. Must be one of \code{"normalized"} or \code{"enriched"}. Default is \code{"normalized"}.}

\item{geom_type}{A character string defining the geometry type for the plot. Can be
either \code{"line"} (line plot) or \code{"col"} (column/bar plot). Default is \code{"line"}.}

\item{filter_cds_divided_by3}{Wthether filter genes whose cds region length can be
divided with 3, default True.}

\item{merge_rep}{Logical; if \code{TRUE}, data from replicates are merged into one plot. Default is \code{TRUE}.}

\item{collapse}{Logical; if \code{TRUE}, replicates or sample groups are collapsed into a single plot. Default is \code{FALSE}.}

\item{sample_color}{A named vector of colors for samples. Default is \code{NULL}, which uses ggplot's default color palette.}

\item{average_smooth}{Logical; if \code{TRUE}, a sliding window average is applied to the ratio data. Default is \code{TRUE}.}

\item{window_size}{Integer specifying the window size for the sliding average. Default is \code{7}.}

\item{mode}{A character string specifying the unit for plotting. Can be either \code{"codon"} or \code{"nt"} (nucleotide). Default is \code{"codon"}.}

\item{add_frame_anno}{Logical; if \code{TRUE}, frame annotations are added to the plot for nucleotide plots. Default is \code{FALSE}.}

\item{frame_relheight}{Numeric value specifying the relative height of the frame annotation plot. Default is \code{0.1}.}

\item{seq_mark}{A character string specifying the type of codon to annotate. Must be one of \code{"both"}, \code{"start"}, or \code{"stop"}. Default is \code{"both"}.}

\item{cds_region_anno}{Optional data frame containing coding sequence annotation.
Default is \code{NULL}.}

\item{cds_col}{A character string specifying the color for visualizing CDS regions. Default is \code{"grey"}.}

\item{cds_anno_col}{A character string specifying the color for coding sequence annotation. Default is \code{"#996600"}.}

\item{cds_region_width}{Numeric value specifying the width of the CDS structure annotation. Default is \code{5}.}

\item{cds_region_scale_x}{Numeric value for scaling the coding sequence on the x-axis. Default is \code{NULL}.}

\item{geom_line_list}{A list of ggplot2 aesthetic modifications for the line geometry. Default is an empty list.}

\item{geom_col_list}{A list of ggplot2 aesthetic modifications for the column geometry. Default is an empty list.}

\item{facet_grid_list}{A list of parameters for modifying `ggplot2::facet_grid()` layout. Default is an empty list.}

\item{...}{Further arguments to be passed to internal functions.}
}
\value{
Returns a \code{ggplot} object or a combined plot list if multiple genes are plotted.
}
\description{
This function creates enrichment track plots to visualize ribosome profiling data.
Users can visualize either ribosome footprint density or the enrichment ratio
(IP/Total) across transcripts. The function generates highly customizable plots,
including options for codon or nucleotide-level resolution, plotting as line or
column, and more.
}
\details{
The function visualizes ribosome profiling data by plotting it as either ribosome footprint
density (from \code{object@normalized.data}) or enrichment ratio (from \code{object@enriched.data}.

Users can choose to display data either by codons or nucleotide, select start and stop codons for annotation,
apply optional smoothing using a sliding window average, and add customizable gene structure annotations.

If `add_frame_anno = TRUE`, frame annotations will be added for a nucleotide plot.
}
\examples{
\dontrun{
data <- ribosomeObj # Assuming ribosomeObj is an object of class ribosomeObj
p <- enrichment_track_plot(
  object = data,
  select_gene = "GeneA",
  select_sample = c("Sample1", "Sample2"),
  assay = "normalized",
  geom_type = "line",
  merge_rep = TRUE
)
print(p) # Visualize the plot
}

}
