% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_sequence.R
\name{fetch_sequence}
\alias{fetch_sequence}
\alias{fetch_sequence,ribosomeObj-method}
\title{FetchSequence}
\usage{
fetch_sequence(
  object,
  output_prefix = NULL,
  pythonPath = NULL,
  type = c("cds", "exon"),
  coding_type = c("NT", "AA"),
  table = 1,
  exclude_stop = c("yes", "no"),
  ...
)

\S4method{fetch_sequence}{ribosomeObj}(
  object,
  output_prefix = NULL,
  pythonPath = NULL,
  type = c("cds", "exon"),
  coding_type = c("NT", "AA"),
  table = 1,
  exclude_stop = c("yes", "no"),
  ...
)
}
\arguments{
\item{object}{ribosomeObj object.}

\item{output_prefix}{A character string specifying the path to the output file.}

\item{pythonPath}{A character string specifying the path to the Python interpreter.}

\item{type}{A character string specifying the type of sequence to fetch ('cds' or 'exon').}

\item{coding_type}{A character string specifying the coding type ('NT' or 'AA').}

\item{table}{Translation table, for which you can give an
NCBI genetic code number or name(https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi),
default 1.}

\item{exclude_stop}{Whether stop to extract amino acid when encounter stop codon.
c("yes","no"), default "yes".}

\item{...}{Useless args.}
}
\value{
None.
}
\description{
FetchSequence

method for fetch_sequence
}
