% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FetchSequence.R
\name{FetchSequence}
\alias{FetchSequence}
\title{FetchSequence}
\usage{
FetchSequence(
  gene_file = NULL,
  genome_file = NULL,
  output_file = NULL,
  type = c("cds", "exon"),
  coding_type = c("NT", "AA"),
  pythonPath = NULL,
  table = 1,
  exclude_stop = c("yes", "no")
)
}
\arguments{
\item{gene_file}{A character string specifying the path to the gene annotation file.}

\item{genome_file}{A character string specifying the path to the reference genome file.}

\item{output_file}{A character string specifying the path to the output file.}

\item{type}{A character string specifying the type of sequence to fetch ('cds' or 'exon').}

\item{coding_type}{A character string specifying the coding type ('NT' or 'AA').}

\item{pythonPath}{A character string specifying the path to the Python interpreter.}

\item{table}{Translation table, for which you can give an
NCBI genetic code number or name(https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi),
default 1.}

\item{exclude_stop}{Whether stop to extract amino acid when encounter stop codon.
c("yes","no"), default "yes".}
}
\value{
None.
}
\description{
FetchSequence
}
