% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneAnnotationExtend.R
\name{gene_anno2Ribominer}
\alias{gene_anno2Ribominer}
\title{gene_anno2Ribominer}
\usage{
gene_anno2Ribominer(longest_trans_file = NULL, output_file = NULL)
}
\arguments{
\item{longest_trans_file}{A character string specifying the path to the file
containing the longest transcript information.}

\item{output_file}{A character string specifying the path to the output file.
If not provided, the function returns the processed data frame.}
}
\value{
A data frame with processed gene annotation information in the format
required by RiboMiner.
}
\description{
gene_anno2Ribominer
}
