% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation_extend.R
\name{gene_anno_extend}
\alias{gene_anno_extend}
\title{gene_anno_extend}
\usage{
gene_anno_extend(
  longest_trans_file = NULL,
  upstream_scale = NULL,
  upstream_extend = NULL,
  downstream_scale = NULL,
  downstream_extend = NULL,
  output_file = NULL
)
}
\arguments{
\item{longest_trans_file}{A character string specifying the path to the file
containing the longest transcript information.}

\item{upstream_scale}{An integer specifying the scale for extending the 5' UTR
region upstream. If not provided, the 5' UTR length is used.}

\item{upstream_extend}{An integer specifying the number of nucleotides to extend
the 5' UTR region upstream. If not provided, the 5' UTR length is used.}

\item{downstream_scale}{An integer specifying the scale for extending the 3' UTR
region downstream. If not provided, the 3' UTR length is used.}

\item{downstream_extend}{An integer specifying the number of nucleotides to extend
the 3' UTR region downstream. If not provided, the 3' UTR length is used.}

\item{output_file}{A character string specifying the path to the output file.
If not provided, the function returns the extended data frame.}
}
\value{
A data frame with extended gene annotation information.
}
\description{
gene_anno_extend
}
