% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bamfiles.R
\name{get_counts_from_bam}
\alias{get_counts_from_bam}
\title{Calculate read counts from BAM files using a GTF annotation}
\usage{
get_counts_from_bam(
  gtf_file = NULL,
  bam_files = NULL,
  type = c("ribo", "rna"),
  isPairedEnd = FALSE,
  exclude_upstream_nt = 60,
  exclude_downstream_nt = 60
)
}
\arguments{
\item{gtf_file}{A character string representing the path to a GTF file for annotation.}

\item{bam_files}{A character vector of paths to BAM files.}

\item{type}{A character string specifying the type of data: "ribo" for ribosome profiling or "rna" for RNA sequencing.}

\item{isPairedEnd}{A logical value indicating whether the BAM files contain paired-end reads. Default is \code{FALSE}.}

\item{exclude_upstream_nt}{An integer specifying the number of nucleotides to exclude upstream for ribosome profiling. Default is 60.}

\item{exclude_downstream_nt}{An integer specifying the number of nucleotides to exclude downstream for ribosome profiling. Default is 60.}
}
\value{
A matrix of read counts with genes as rows and samples as columns.
}
\description{
This function processes BAM files to calculate read counts for specified features
using a GTF file, with options for ribosomal profiling and RNA sequencing.
}
\details{
For ribosome profiling, the function adjusts the CDS regions by excluding specified nucleotides upstream and downstream before calculating counts.
}
\examples{
\dontrun{
gtf_file <- "path/to/annotation.gtf"
bam_files <- c("sample1.bam", "sample2.bam")
counts <- get_counts_from_bam(gtf_file, bam_files, type = "rna", isPairedEnd = TRUE)
}
}
