% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nomalized_counts.R
\name{get_nomalized_counts}
\alias{get_nomalized_counts}
\alias{get_nomalized_counts,ribosomeObj-method}
\title{Get Normalized Counts for Transcript Positions}
\usage{
get_nomalized_counts(
  object,
  exclude_upstreamCodon = 5,
  exclude_downstreamCodon = 5,
  ...
)

\S4method{get_nomalized_counts}{ribosomeObj}(
  object,
  exclude_upstreamCodon = 5,
  exclude_downstreamCodon = 5,
  ...
)
}
\arguments{
\item{object}{ribosomeObj object.}

\item{exclude_upstreamCodon}{An integer specifying the number of codons to exclude from the start of the coding sequence.
Default is 5}

\item{exclude_downstreamCodon}{An integer specifying the number of codons to exclude from the end of the coding sequence.
       Default is 5.
The nomalization methods for ribosome density. "average" is calculated by
the count at each position divided by mean density across cds region. "rpm"
is calculated by the count at each position divided by the total counts and multiplied with 10^6.
Default is "average".}

\item{...}{Useless args.}
}
\value{
A data frame with normalized counts for each position in each transcript across all samples.
        The returned data frame includes all original columns from qc_df plus a new column 'norm_exp'
        representing the normalized expression.
}
\description{
This function normalizes the counts at individual positions to the average counts of the transcript,
with options to exclude a specified number of codons from the start and end of the coding sequence.
}
\details{
The function performs the following steps:
  1. Reads and processes the gene annotation file.
  2. Calculates total reads and average expression for each transcript, excluding specified regions.
  3. Normalizes the counts at each position by dividing by the average expression of the transcript.
}
