% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboseqTools_implementation.R
\name{get_serp_plot_df}
\alias{get_serp_plot_df}
\title{Plot a profile along a gene}
\usage{
get_serp_plot_df(
  object = NULL,
  type = c("enrichment", "rpm"),
  gene_name = NULL,
  samples = NULL,
  sample1 = NULL,
  sample2 = NULL,
  exp,
  rep,
  bin = c("byaa", "bynuc"),
  window_size = 45,
  conf.level = 0.95
)
}
\arguments{
\item{object}{A \code{serp_data} object. Must contain raw (unnormalized) read counts.}

\item{type}{Plot type. One of \code{enrichment}, \code{rpm}.}

\item{gene_name}{Name of the gene/ORF to plot.}

\item{samples}{Samples to plot. If missing, all samples will be plotted.}

\item{sample1}{Name of the first sample (the numerator). If missing, the default sample1 of the data set
will be used.}

\item{sample2}{Name of the second sample (the denominator). If missing, the default sample2 of the data set
will be used.}

\item{exp}{Character vector of experiments to plot. If missing, all experiments are plotted.}

\item{rep}{Character vector of replicates to plot. If missing, all replicates will be plotted.}

\item{bin}{Bin level (\code{bynuc} or \code{byaa}). If missing, the default binning level of the data set
will be used.}

\item{window_size}{Neighborhood size for the confidence interval calculation in nucleotides. If missing, the default
window size of the data set will be used.}

\item{conf.level}{Confidence level.}
}
\value{
A data.frame.
}
\description{
For each position, a confidence interval is plotted. Transparency reflects the total number of reads
contributing to the confidence interval.
}
