% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generic_methods.R
\name{get_track_df}
\alias{get_track_df}
\alias{get_track_df,ribosomeObj-method}
\title{Generate Track Data Frame for Selected Genes}
\usage{
get_track_df(
  object,
  select_gene = NULL,
  select_sample = NULL,
  unit_to_codon = FALSE,
  ...
)

\S4method{get_track_df}{ribosomeObj}(
  object,
  select_gene = NULL,
  select_sample = NULL,
  unit_to_codon = FALSE,
  ...
)
}
\arguments{
\item{object}{ribosomeObj object.}

\item{select_gene}{A character vector of gene names to filter the gene annotation data.
If NULL, no filtering is applied. Default is NULL.}

\item{select_sample}{A character vector of sample names to filter. Default is NULL.}

\item{unit_to_codon}{Whether transform nuceotide position to codon position. Default is FALSE.}

\item{...}{Useless args.}
}
\value{
A data frame containing the filtered and transformed data suitable for
tracking gene expression. The returned data frame includes columns for sample,
gene name, transcript ID, transcript position (transpos), normalized expression (density),
and a fixed column (type) set to 'ribo'.
}
\description{
This function extracts and prepares track data for selected genes from gene
annotation and normalized expression files.
It filters gene annotation data for selected genes, merges it with normalized
expression data, and formats it for further analysis or visualization.
}
\examples{
\dontrun{# Assuming `longest_trans_file` and `normed_file` are already defined:
track_df <- get_track_df(
  object = ribosomeObj,
  select_gene = c("Gene1", "Gene2")
)}

}
