% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seRPdataAnalysis.R
\name{load_enrich_data}
\alias{load_enrich_data}
\title{Load and enrich ribosome profiling data}
\usage{
load_enrich_data(
  data_type = c("raw", "sw"),
  enrich_data = NULL,
  gene_names = NULL,
  sample_name = NULL
)
}
\arguments{
\item{data_type}{The type of input data: "raw" or "sw".}

\item{enrich_data}{A character vector of file paths to the input data files.}

\item{gene_names}{A character vector of gene names to filter on.}

\item{sample_name}{A character vector of sample names to use instead of file names.}
}
\value{
A data frame with columns for gene_name, trans_id, codon_pos, type,
sample, density, and (density_sd).
}
\description{
This function processes raw or smoothed ribosome profiling data and returns
a data frame with mean density and standard deviation values for each codon
position, gene, sample, and type (ribosome profiling).
}
\examples{
\dontrun{
load_enrich_data(data_type = "raw", enrich_data = c("sample1.txt", "sample2.txt"),
                 gene_names = c("GAPDH", "ACTB"))
load_enrich_data(data_type = "sw", enrich_data = c("sample1.txt", "sample2.txt"),
                 sample_name = c("Sample A", "Sample B"))
}

}
