% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{load_expression_data}
\alias{load_expression_data}
\title{Load expression data from text files}
\usage{
load_expression_data(sample_name = NULL)
}
\arguments{
\item{sample_name}{A character vector containing sample names. If NULL, the
function will use the file names as sample names.}
}
\value{
A list containing two matrices:
\describe{
  \item{count_matrix}{A matrix of gene expression counts for all samples.}
  \item{tpm_matrix}{A matrix of gene expression TPM values for all samples.}
}
}
\description{
This function loads gene expression data from text files in a specified
directory. It returns both count and TPM (transcripts per million) matrices
for all samples.
}
