% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{load_metagene_data}
\alias{load_metagene_data}
\title{Load Metagene Data}
\usage{
load_metagene_data(sample_name = NULL, group_name = NULL)
}
\arguments{
\item{sample_name}{A character vector specifying the names of the input files.
If NULL, all files in the directory will be used.}

\item{group_name}{A character vector specifying the group names for each set
of input files. If NULL, no grouping will be performed.}
}
\value{
A data frame containing the loaded metagene data with columns for
position, density, sample name, and group name.
}
\description{
This function loads metagene data from text files in the specified directory.
}
\examples{
\dontrun{
load_metagene_data(sample_name = c("sample1.txt", "sample2.txt"))
}

}
