% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{load_qc_data}
\alias{load_qc_data}
\title{load_qc_data function}
\usage{
load_qc_data(sample_name = NULL, group_name = NULL)
}
\arguments{
\item{sample_name}{A vector of file names for the samples to be loaded.}

\item{group_name}{The name of the group to which the loaded samples belong.}
}
\value{
Returns a merged data frame containing QC data for all loaded samples.
}
\description{
A function to load quality control (QC) data from files and create a merged
data frame.
}
