% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_gene_plot.R
\name{load_rna_coverage}
\alias{load_rna_coverage}
\title{Load RNA Coverage Data}
\usage{
load_rna_coverage(
  trans_density_file = NULL,
  sample_name = NULL,
  replicate = NULL,
  group = NULL
)
}
\arguments{
\item{trans_density_file}{A character vector of file names that contain RNA transcript density data.
Each file should be in a tab-delimited format.}

\item{sample_name}{A character vector of sample names corresponding to each file in `trans_density_file`.}

\item{replicate}{A character vector indicating replicate information for each file in `trans_density_file`.
Defaults to `NULL`, in which case replicate information is set to `NA`.}

\item{group}{A character vector indicating the group to which each sample belongs.
Defaults to `NULL`, in which case group information is set to `NA`.}
}
\value{
A data frame containing the processed RNA coverage data, with columns for transcript ID (`trans_id`),
transcript position (`trans_pos`), reads per million (`RPM`), sample name (`sample`), replicate information (`rep`),
group information (`group`), and a type column (`type`) indicating the data type as RNA.
}
\description{
This function loads RNA coverage data from a specified file or files, processes it,
and returns a data frame with additional metadata including sample name, replicate, and group information.
}
\details{
The function reads in tab-delimited RNA transcript density data from the specified files, filters
out rows where RPM is zero, and adds metadata columns such as sample name, replicate, and group. The resulting
data frame is created by binding the processed data from all input files together.
}
