% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboseqTools_implementation.R
\name{load_selective_ribosome}
\alias{load_selective_ribosome}
\title{load selective ribosome data}
\usage{
load_selective_ribosome(
  ...,
  object = NULL,
  normalize = FALSE,
  bin = c("bynuc", "byaa"),
  exclude = list(),
  defaults = list()
)
}
\arguments{
\item{...}{Name-value pairs of lists. The name of each argument will be the name of an experiment.
The name of each element will be the sample type (e.g. TT for total translatome), the value
of each element must be a character vector of file paths, where each file is a read count
table of one replicate experiment. Replicate order must match between sample types.}

\item{object}{object ribosomeObj Object.}

\item{normalize}{Normalize the read counts to library size? Output will then be in RPM.}

\item{bin}{Bin the data. \code{bynuc}: No binning (i.e. counts per nucleotide). \code{byaa}: Bin by residue.}

\item{exclude}{Genes to exclude in all future analyses. This genes will also be excluded from total read count
calculation. Note that the raw count tables will not be modified. Named list with names corresponding to
experiments. If a character vector of gene names is given, these genes will be excluded from all
experiments.}

\item{defaults}{Default parameters of the data set.}
}
\value{
An object of class \code{serp_data}
}
\description{
load selective ribosome data
}
