% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{load_track_data}
\alias{load_track_data}
\title{Load gene expression data from ribo and rna files}
\usage{
load_track_data(
  ribo_file = NULL,
  rna_file = NULL,
  sample_name = NULL,
  gene_list = NULL
)
}
\arguments{
\item{ribo_file}{A character vector specifying the names of ribo file(s).}

\item{rna_file}{A character vector specifying the names of rna file(s).}

\item{sample_name}{A character vector specifying the name(s) of the sample(s).}

\item{gene_list}{A character vector specifying the gene names to be included.}
}
\value{
A data frame containing gene expression data from both ribo and rna
files, filtered by the specified gene list and merged into a single data frame.
}
\description{
This function reads in gene expression data from ribo and rna files, filters
by a given gene list, and merges them into a single data frame. The resulting
data frame includes columns for gene name, transcript ID, position, density,
sample name, and type of expression data (ribo or rna).
}
\examples{
\dontrun{
load_track_data(ribo_file = c("file1_ribo.txt", "file2_ribo.txt"),
                rna_file = c("file1_rna.txt", "file2_rna.txt"),
                sample_name = c("sample1", "sample2","sample1", "sample2"),
                gene_list = c("gene1", "gene2"))
}

}
