% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagene_plot.R
\name{metagene_plot}
\alias{metagene_plot}
\title{Metagene Plot Generation}
\usage{
metagene_plot(
  longest_trans_file = NULL,
  normed_file = NULL,
  min_counts = 64,
  min_cds_length = 600,
  relative_distance = c(-25, 500),
  type = c("st", "sp"),
  mode = c("nt", "codon"),
  collapse = FALSE,
  geom_line_params = list(),
  facet_wrap_params = list()
)
}
\arguments{
\item{longest_trans_file}{A file path to the longest transcript data, containing gene annotations.
Default is NULL.}

\item{normed_file}{A file path to the normalized gene expression data. Default is NULL.}

\item{min_counts}{The minimum count threshold to include a transcript in the analysis.
Default is 64.}

\item{min_cds_length}{The minimum CDS length required to include a transcript. Default is 600.}

\item{relative_distance}{A numeric vector specifying the relative distance range from the start
or stop codon to plot. Default is c(-25, 500).}

\item{type}{Character vector, either "st" for start codon or "sp" for stop codon, to specify the
reference point for relative distances. Default is c("st", "sp") and will be matched in order.}

\item{mode}{Character vector, either "nt" for nucleotide or "codon" for codon resolution in the plot.
Default is c("nt", "codon") and will be matched in order.}

\item{collapse}{Logical, whether to collapse data across samples or not. Default is FALSE.}

\item{geom_line_params}{List of additional parameters for geom_line, used to customize the line elements
of the plot. Default is an empty list().}

\item{facet_wrap_params}{List of additional parameters for facet_wrap, used to customize the faceting
of the plot. Default is an empty list().}
}
\value{
A ggplot object representing the metagene analysis plot.
}
\description{
This function generates a metagene plot based on normalized expression data across transcripts.
It allows for flexible filtering, normalization, and plotting of gene expression data relative
to translation start (st) or stop (sp) sites. Data can be plotted at nucleotide or codon resolution.
}
\examples{
\dontrun{# Assuming `longest_trans_file` and `normed_file` are paths to your data files:
metagene_plot(
  longest_trans_file = "path/to/longest_transcripts.tsv",
  normed_file = "path/to/normed_expression.tsv",
  min_counts = 64,
  min_cds_length = 600,
  relative_distance = c(-25, 500),
  type = "st",
  mode = "nt"
)}

}
