% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboseqTools_implementation.R
\name{metagene_profile}
\alias{metagene_profile}
\title{Generate Metagene Profile Plot}
\usage{
metagene_profile(
  df = NULL,
  ytrans = "log2",
  align = "start",
  aes_col = NULL,
  facet_grid_params = list(),
  conf.level = 0.95,
  ci.alpha = 0.3
)
}
\arguments{
\item{df}{A dataframe containing positional metagene data. Should include columns
like `pos` (position), `summary` (signal summary), and a grouping variable for fill/color.}

\item{ytrans}{A string specifying the transformation for the y-axis. Default is `"log2"`.
Example values are `"log10"`, `"sqrt"` or `"identity"` for no transformation.}

\item{align}{A string indicating how the metagene data is aligned. Can be `"start"`
for start codon or `"stop"` for stop codon. Used in formatting the x-axis label. Default is `"start"`.}

\item{aes_col}{Aesthetic mapping for the fill and color of the plot. This variable
represents the grouping factor for different metagene profiles. Should match a column name in `df`.}

\item{facet_grid_params}{A list of parameters to pass to `facet_grid()`, allowing for faceted plots by different factors (e.g., experimental replicates).
Default is an empty list.}

\item{conf.level}{The confidence level for the bootstrap confidence intervals (CI). Must be a number between `0` and `1`. Default is `0.95`.}

\item{ci.alpha}{The alpha transparency value (between `0` and `1`) for the confidence interval shade. Default is `0.3`.}
}
\value{
A ggplot object visualizing the metagene profile with confidence intervals.
}
\description{
This function creates a metagene profile plot using `ggplot2`, visualizing
ribosome profiling data or other biological signals aligned to a specific reference
point such as the start or stop codon. It also supports confidence intervals (CI)
for bootstraps and faceting for additional dimensions of group or experiment.
}
