% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripeptide_motif_score.R
\name{peptide_motif_score2}
\alias{peptide_motif_score2}
\alias{peptide_motif_score2,ribosomeObj-method}
\title{Peptide Motif Scoring Function}
\usage{
peptide_motif_score2(
  object,
  norm_type = c("rpm", "average"),
  window = 50,
  occurrence_threshold = 100,
  ...
)

\S4method{peptide_motif_score2}{ribosomeObj}(
  object,
  norm_type = c("rpm", "average"),
  window = 50,
  occurrence_threshold = 100,
  ...
)
}
\arguments{
\item{object}{ribosomeObj object.}

\item{norm_type}{A character string specifying the normalization type. It can be one of
"rpm" or "average".
Default is "normed_count".}

\item{window}{An integer specifying the window size for motif scanning.
Default is 50.}

\item{occurrence_threshold}{An integer specifying the threshold for motif occurrences.
Default is 100.}

\item{...}{Useless args.}
}
\value{
The function creates output files with motif scores and prints a message for each processed sample.
}
\description{
This function calculates the motif scores for peptides based on provided files.
}
\examples{
\dontrun{# Example usage:
peptide_motif_score2(object = ribosomeObj)}

}
