% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodicity_check.R
\name{periodicity_check}
\alias{periodicity_check}
\alias{periodicity_check,ribosomeObj-method}
\title{Check Periodicity in Ribosome Profiling Data}
\usage{
periodicity_check(
  object,
  read_length = c(27, 30),
  relative_dist = c(0, 150),
  period_max = 10,
  ...
)

\S4method{periodicity_check}{ribosomeObj}(
  object,
  read_length = c(27, 30),
  relative_dist = c(0, 150),
  period_max = 10,
  ...
)
}
\arguments{
\item{object}{A `ribosomeObj` object containing the raw counts and metadata.}

\item{read_length}{A numeric vector of length 2 specifying the range of read lengths to consider.
Default is c(27, 30).}

\item{relative_dist}{A numeric vector of length 2 indicating the distance range around the start codon to evaluate.
Default is c(0, 150).}

\item{period_max}{A numeric value indicating the maximum period (in nucleotides) to display in the plot.
Default is 10.}

\item{...}{Additional arguments to be passed to or from other methods.}
}
\value{
A ggplot object displaying the amplitude of different periods across samples and read lengths.
}
\description{
This function performs a periodicity check on reads from ribosome profiling data.
It uses the Discrete Fourier Transform to evaluate the periodicity of ribosomal reads
around start codons, and visualizes the result.
}
