% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_mapinfo}
\alias{plot_mapinfo}
\title{Plot mapping information from mapinfo files}
\usage{
plot_mapinfo(
  mapinfo_file = NULL,
  file_name = NULL,
  plot_type = c("barplot", "table"),
  geom_col_params = list()
)
}
\arguments{
\item{mapinfo_file}{A character vector of file paths to the mapinfo files.}

\item{file_name}{An optional character vector of names to assign to the samples.
If NULL, the sample name will be extracted from the mapinfo file paths.}

\item{plot_type}{A string indicating the type of plot to generate.
Either "barplot" for a stacked barplot of mapping percentages, or "table" for
a table of mapping counts.}

\item{geom_col_params}{A list of parameters to pass to the \code{geom_col} function.
Useful for modifying the appearance of the barplot.}
}
\value{
A barplot or table depending on the \code{plot_type} parameter.
}
\description{
This function reads in mapinfo files and extracts mapping information for each file.
The extracted information is then plotted either as a barplot or a table.
}
\examples{
\dontrun{
# Generate a barplot of mapping percentages
plot_mapinfo(mapinfo_file = c("sample1_mapinfo.txt", "sample2_mapinfo.txt"), plot_type = "barplot")

# Generate a table of mapping counts
plot_mapinfo(mapinfo_file = c("sample1_mapinfo.txt", "sample2_mapinfo.txt"), plot_type = "table")
}

}
