% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_polarity_sore.R
\name{polarity_score_plot}
\alias{polarity_score_plot}
\title{Plot Polarity Scores}
\usage{
polarity_score_plot(
  polarity_score_df = NULL,
  aes_col_var = c("sample", "group"),
  linewidth = 1,
  orf_col = "grey",
  orf_pos = -0.2,
  orf_width = 4,
  text_pos = 0.1,
  text_size = 4,
  facet = FALSE,
  scales = "free",
  ncol = 2
)
}
\arguments{
\item{polarity_score_df}{A data frame containing the polarity scores.}

\item{aes_col_var}{A character string specifying the column variable for the
aesthetic mapping. It can be either 'sample' or 'group'. Default is 'sample'.}

\item{linewidth}{A numeric value specifying the width of the density lines.
Default is 1.}

\item{orf_col}{A character string specifying the color of the ORF line.
Default is "grey".}

\item{orf_pos}{A numeric value specifying the position of the ORF line.
Default is -0.2.}

\item{orf_width}{A numeric value specifying the width of the ORF line.
Default is 4.}

\item{text_pos}{A numeric value specifying the position of the text label.
Default is 0.1.}

\item{text_size}{A numeric value specifying the size of the text label.
Default is 4.}

\item{facet}{A logical value specifying whether to facet the plot by group.
Default is FALSE.}

\item{scales}{A character string specifying the scales for faceting.
Default is "free".}

\item{ncol}{A numeric value specifying the number of columns for faceting.
Default is 2.}
}
\value{
A ggplot2 object representing the polarity score plot.
}
\description{
This function creates a plot of polarity scores for genes based on RNA-seq data.
The function takes a data frame containing polarity scores and optionally a
grouping factor. It then creates a density plot of the polarity scores, with
the option to facet the plot by group.
}
