% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{pre_count_tpm_data}
\alias{pre_count_tpm_data}
\title{Preprocess SAM files to calculate count and TPM data}
\usage{
pre_count_tpm_data(sam_file = NULL, out_file = NULL, type = c("ribo", "rna"))
}
\arguments{
\item{sam_file}{A character vector of SAM file paths.}

\item{out_file}{A character vector of output file names.}

\item{type}{Specifies the type of the input file. Default is 'ribo'.}
}
\description{
This function takes SAM files as input, processes it using JuliaCall package
and generates count and TPM data.
}
\examples{
\dontrun{
pre_count_tpm_data(sam_file = c("path/to/samfile1.sam","path/to/samfile2.sam"),
                   out_file = c("output1.txt","output2.txt"),
                   type = "rna")
}

}
