% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seRPdataAnalysis.R
\name{pre_enrichment_data}
\alias{pre_enrichment_data}
\title{pre_enrichment_data function}
\usage{
pre_enrichment_data(
  riboIP_file = NULL,
  riboInput_file = NULL,
  output_file = NULL,
  pythonPath = NULL
)
}
\arguments{
\item{riboIP_file}{A character vector containing the file path(s) to the
ribosome profiling IP sample(s).}

\item{riboInput_file}{A character vector containing the file path(s) to the
ribosome profiling input control sample(s).}

\item{output_file}{A character vector containing the desired output file
path(s) for each sample's enrichment analysis results.}

\item{pythonPath}{An optional parameter specifying the file path to a specific
Python installation to use (if not already set as the default in reticulate).}
}
\description{
This function performs enrichment analysis on ribosome profiling data.
}
\examples{
\dontrun{
pre_enrichment_data(riboIP_file = "path/to/IP_sample.bam",
                    riboInput_file = "path/to/Input_sample.bam",
                    output_file = "path/to/output_file.txt")
}

}
