% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{pre_gene_trans_density}
\alias{pre_gene_trans_density}
\title{Calculate density data for a given gene annotation file}
\usage{
pre_gene_trans_density(gene_anno = NULL, density_file = NULL, out_file = NULL)
}
\arguments{
\item{gene_anno}{A gene annotation file in the format accepted by XAM
(default is NULL).}

\item{density_file}{A character vector of input file names for which to
calculate density data (default is NULL).}

\item{out_file}{A character vector of output file names corresponding to each
input file (default is NULL). The length of this vector must be equal to the
length of \code{density_file}.}
}
\value{
This function does not return anything; it writes the output files to
disk.
}
\description{
This function calculates density data for a given gene annotation file using
the XAM package in Julia. It writes the output files to the "2.density-data"
directory.
}
\examples{
\dontrun{
# Calculate density data for a single input file
pre_gene_trans_density(gene_anno = "gene_annotation.txt",
                       density_file = "input_file.txt",
                       out_file = "output_file.txt")

# Calculate density data for multiple input files
pre_gene_trans_density(gene_anno = "gene_annotation.txt",
                       density_file = c("input_file1.txt", "input_file2.txt"),
                       out_file = c("output_file1.txt", "output_file2.txt"))
}

}
