% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{pre_longest_trans_info}
\alias{pre_longest_trans_info}
\title{Get information about longest transcripts in a GTF file}
\usage{
pre_longest_trans_info(gtf_file = NULL, out_file = NULL, pythonPath = NULL)
}
\arguments{
\item{gtf_file}{path to the GTF file, ".gz" compressed format is acceptted.}

\item{out_file}{path to the output file.}

\item{pythonPath}{path to the Python executable (optional).}
}
\value{
This function does not return anything. It writes the output to a file.
}
\description{
This function uses a Python script to extract information about the longest
transcripts in a GTF file. The script generates an output file containing the
transcript ID, gene ID,... and length of the longest transcript for each gene.
}
\examples{
\dontrun{
pre_longest_trans_info(gtf_file = "my_gtf_file.gtf",
                       out_file = "longest_transcripts.txt",
                       pythonPath = "/usr/bin/python3")
}

}
