% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{pre_metagene_data}
\alias{pre_metagene_data}
\title{Calculate Metagene Data}
\usage{
pre_metagene_data(
  mapping_type = c("genome", "transcriptome"),
  gene_anno = NULL,
  density_file = NULL,
  out_file = NULL,
  mode = c("st", "sp"),
  type = "codon",
  cdslength = 600,
  expression = 30,
  exclude = 90
)
}
\arguments{
\item{mapping_type}{The mapping type for your sam files, "genome" or "transcriptome".}

\item{gene_anno}{A data.frame containing gene annotation information which is
needed when mapping_type = "genome".}

\item{density_file}{A character vector of file names for ribosome density data.}

\item{out_file}{A character vector of output file names for metagene data.}

\item{mode}{A character indicating the mode of analysis, "st" or "sp".
Default is "st".}

\item{type}{A character indicating the type of analysis, "nt" or "codon". Default is "codon".}

\item{cdslength}{An integer indicating the length of CDS. Default is 600.}

\item{expression}{An integer indicating the minimum expression value. Default is 30.}

\item{exclude}{An integer indicating the number of nucleotides to exclude from
both ends of the coding region. Default is 90.}
}
\description{
This function calculates metagene data using ribosome density files and gene
annotation.
}
\examples{
# Assume that we have a gene annotation data frame called 'gene_annot' and two ribosome density
# files called 'file1.bam' and 'file2.bam' in '2.density-data/' directory. The following code
# will calculate the metagene data and save the results in '3.metagene-data/' directory with
# names 'metagene1.tsv' and 'metagene2.tsv':
#
# pre_metagene_data(gene_anno = gene_annot,
#                   density_file = c("file1.bam", "file2.bam"),
#                   out_file = c("metagene1.txt", "metagene2.txt"))

}
