% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribosome_data_preparing.R
\name{pre_qc_data}
\alias{pre_qc_data}
\title{pre_qc_data function}
\usage{
pre_qc_data(
  mapping_type = c("genome", "transcriptome"),
  julia_path = NULL,
  XAM_version = c("3", "4"),
  longest_trans_file = NULL,
  sam_file = NULL,
  bam_file = NULL,
  out_file_prefix = NULL,
  rep_name = NULL,
  group_name = NULL,
  has_created_data = FALSE,
  assignType = c("end5", "end3"),
  seq_type = c("pairedEnd", "singleEnd")
)
}
\arguments{
\item{mapping_type}{The mapping type for your sam files, "genome" or "transcriptome".}

\item{julia_path}{The julia program path on the computer.}

\item{XAM_version}{The XAM version you have installed, "3"(0.3.1) or "4"(0.4.0).}

\item{longest_trans_file}{A string specifying the path to the longest transcript
file.}

\item{sam_file}{A character vector specifying the paths to the SAM files.}

\item{bam_file}{A character vector specifying the paths to the BAM files.}

\item{out_file_prefix}{A character string specifying the prefix for the output files generated during
the QC analysis.}

\item{rep_name}{The replicates name. Default NULL.}

\item{group_name}{The group name. Default NULL.}

\item{has_created_data}{whether you have created raw data and save time to renalysis.
Defaluts FALSE.}

\item{assignType}{The reads assignment type, 5'end or 3'end, choices c("end5","end3"). Default 5'end.}

\item{seq_type}{The sequencing type for fastq files, "singleEnd" or "pairedEnd".}
}
\value{
The function does not return a value, but outputs QC results in text
files.
}
\description{
This function performs quality control (QC) analysis on ribosome profiling
data. The input is a SAM/BAM file generated from ribosome profiling data and the
output is a QC result in a text file format.
}
\examples{
\dontrun{
pre_qc_data(longest_trans_file = "path/to/longestTransInfo.txt",
            sam_file = c("path/to/sample1.sam", "path/to/sample2.sam"),
            out_file = c("path/to/sample1_QC_result.txt", "path/to/sample2_QC_result.txt"))
}

}
