% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcess.R
\name{pre_ribo_density_data}
\alias{pre_ribo_density_data}
\title{Calculate ribosome density data}
\usage{
pre_ribo_density_data(sam_file = NULL, out_file = NULL, min = 23, max = 35)
}
\arguments{
\item{sam_file}{A character vector of SAM file paths.}

\item{out_file}{A character vector of output file names.}

\item{min}{The minimum length of reads to be considered for calculating density
(default is 23).}

\item{max}{The maximum length of reads to be considered for calculating density
(default is 35).}
}
\value{
No explicit return value. Output files are written to "2.density-data" directory.
}
\description{
This function calculates the ribosome density data using XAM package in Julia.
}
\examples{
\dontrun{
# Calculate ribosome density data for a single SAM file
pre_ribo_density_data(sam_file = "sample.sam", out_file = "sample_density.txt")

# Calculate ribosome density data for multiple SAM files
pre_ribo_density_data(sam_file = c("sample1.sam", "sample2.sam"),
                      out_file = c("sample1_density.txt", "sample2_density.txt"))
}

}
