% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{pre_rna_coverage_data}
\alias{pre_rna_coverage_data}
\title{Calculate RNA coverage data}
\usage{
pre_rna_coverage_data(sam_file = NULL, bam_file = NULL, out_file = NULL)
}
\arguments{
\item{sam_file}{A character vector specifying the path(s) of the input SAM file(s).}

\item{bam_file}{If using sam file runs slowly, try using bam file instead.}

\item{out_file}{A character vector specifying the name(s) of the output file(s).
The output file(s) will be saved in the "2.density-data" directory.}
}
\description{
This function generates pre-processing RNA coverage data for a given SAM/BAM file.
}
\examples{
\dontrun{
# Generate pre-processing RNA coverage data for a single SAM file
pre_rna_coverage_data(sam_file = "/path/to/sam_file.sam", out_file = "coverage_data.txt")

# Generate pre-processing RNA coverage data for multiple SAM files
pre_rna_coverage_data(sam_file = c("/path/to/sam_file_1.sam", "/path/to/sam_file_2.sam"),
                       out_file = c("coverage_data_1.txt", "coverage_data_2.txt"))
}

}
