% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seRPdataAnalysis.R
\name{pre_slideWindow_enrichment_data}
\alias{pre_slideWindow_enrichment_data}
\title{Preprocess ribosome profiling data using sliding window enrichment analysis.}
\usage{
pre_slideWindow_enrichment_data(
  riboIP_file = NULL,
  riboInput_file = NULL,
  gene_list = NULL,
  output_file = NULL,
  pythonPath = NULL
)
}
\arguments{
\item{riboIP_file}{A character vector of file paths to the ribosome profiling
IP samples.}

\item{riboInput_file}{A character vector of file paths to the ribosome profiling
Input samples.}

\item{gene_list}{A character vector of file path to the gene list.}

\item{output_file}{A character vector of output file names for SWEA results.}

\item{pythonPath}{A character string specifying the path to the Python executable.
Defaults to NULL, which results in automatic detection of an available Python
installation.}
}
\value{
The function invisibly returns a list of processed files.
}
\description{
This function preprocesses ribosome profiling data by running sliding window
enrichment analysis (SWEA) on the input files.
}
\details{
This function calls a Python script provided with the RiboProfiler
package to perform sliding window enrichment analysis on ribosome profiling data.
}
\examples{
\dontrun{
# Preprocess ribosome profiling data using sliding window enrichment analysis
pre_slideWindow_enrichment_data(riboIP_file = c("IP1.density.txt", "IP2.density.txt"),
                                riboInput_file = c("Input1.density.txt", "Input2.density.txt"),
                                gene_list = c("gene1","gene2"),
                                output_file = c("enrich_IP1.txt", "enrich_IP2.txt"))
}

}
