% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PsiteOffsetCheck.R
\name{PsiteOffsetCheck}
\alias{PsiteOffsetCheck}
\title{PsiteOffsetCheck}
\usage{
PsiteOffsetCheck(
  qc_data = NULL,
  relative_distance = c(-18, 12),
  read_length = c(27, 32),
  label_size = 3,
  label_pos = c(0.9, 0.9),
  line_col = "#993366",
  peak_line_col = "red"
)
}
\arguments{
\item{qc_data}{A data frame containing quality control data.}

\item{relative_distance}{A vector of two values representing the range of relative distances
to filter by. Default is c(-18,12).}

\item{read_length}{A vector of two values representing the range of read lengths to filter by.
Default is c(27,32).}

\item{label_size}{The size of the text labels. Default is 3.}

\item{label_pos}{A vector of two values representing the position of the labels. Default is c(0.9,0.9).}

\item{line_col}{The color of the line. Default is "#993366".}

\item{peak_line_col}{The color of the peak line. Default is "red".}
}
\value{
A list with two elements:
\item{summary_offset}{A data frame containing the summary of offsets.}
\item{plot}{A ggplot object representing the plot of read counts against distance.}
}
\description{
PsiteOffsetCheck
}
