% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{qc_plot}
\alias{qc_plot}
\title{qc_plot function}
\usage{
qc_plot(
  qc_data = NULL,
  type = c("length", "length_frame", "feature", "frame"),
  geom_col_list = list(),
  facet_wrap_list = list()
)
}
\arguments{
\item{qc_data}{The input data frame for QC analysis. Requires columns "group",
"sample", "length", "counts", "framest" and "feature".}

\item{type}{The type of QC plot to create. Must be one of "length",
"length_frame","frame" or "feature".}

\item{geom_col_list}{A list of additional ggplot2 layers to add to the plot
using the \code{geom_col} function.}

\item{facet_wrap_list}{A list of additional ggplot2 layers to add to the plot
using the \code{facet_wrap} function.}
}
\value{
Returns a ggplot2 object representing the created QC plot.
}
\description{
A function to create quality control (QC) plots for read length, frame and
features.
}
\examples{
\dontrun{
qc_plot(qc_data = my_data, type = "length")
}

}
