% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFuncs.R
\name{reference_index_build}
\alias{reference_index_build}
\title{Build HISAT2 index for reference genome}
\usage{
reference_index_build(
  reference_file = NULL,
  prefix = NULL,
  threads = 8,
  hisat2_build_params = list()
)
}
\arguments{
\item{reference_file}{path to the reference genome file in FASTA format}

\item{prefix}{prefix for the output files}

\item{threads}{number of CPU threads to use for building the index}

\item{hisat2_build_params}{a list of additional parameters to pass to HISAT2
build command}
}
\description{
This function builds a HISAT2 index for the given reference genome using the
HISAT2 tool.
}
\examples{
\dontrun{
# Build a HISAT2 index for human reference genome
reference_index_build(reference_file = "hg38.fa", prefix = "hg38")
}

}
