% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_occupancy_from_motif.R
\name{rel_dist_motif}
\alias{rel_dist_motif}
\alias{rel_dist_motif,ribosomeObj-method}
\title{Calculate Relative Distance of Motifs within Transcripts}
\usage{
rel_dist_motif(
  object,
  min_counts = 64,
  motif = NULL,
  upstream = -50,
  downstream = 50,
  ...
)

\S4method{rel_dist_motif}{ribosomeObj}(
  object,
  min_counts = 64,
  motif = NULL,
  upstream = -50,
  downstream = 50,
  ...
)
}
\arguments{
\item{object}{ribosomeObj object.}

\item{min_counts}{Integer, minimum count threshold for filtering transcripts, default is 64.}

\item{motif}{Character, the specific motif to search for, default is NULL.}

\item{upstream}{Integer, upstream distance to consider relative to motif position, default is -50.}

\item{downstream}{Integer, downstream distance to consider relative to motif position, default is 50.}

\item{...}{Useless args.}
}
\value{
Invisible NULL, as the function primarily writes output to files.
}
\description{
This function calculates the relative distances of specified motifs within transcripts
from normalized data files, filtered by minimum expression counts, and outputs the results.
It handles multiple samples and uses external Python scripts for detailed calculations.
}
\examples{
\dontrun{rel_dist_motif(
               object = ribosomeObj,
               min_counts = 64,
               motif = "PKP",
               upstream = -50,
               downstream = 50)}


}
