% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel_dist_motif.R
\name{rel_dist_motif_plot}
\alias{rel_dist_motif_plot}
\title{Plot Relative Motif Occupancy from Motif Density Files}
\usage{
rel_dist_motif_plot(
  motif_occupancy_file = NULL,
  sample_name = NULL,
  group_name = NULL,
  motif = NULL,
  site = "P"
)
}
\arguments{
\item{motif_occupancy_file}{Character vector, paths to the motif occupancy files.}

\item{sample_name}{Character vector, names of the samples corresponding to the files.}

\item{group_name}{Optional character vector, names of groups for each sample.
If provided, averages density across samples in the same group.}

\item{motif}{Character, the specific motif being analyzed.}

\item{site}{Character, label for the site being considered in the motif analysis,
default is "P" (e.g., for the Peptidyl site in ribosome profiling).}
}
\value{
An ggplot object showing the relative motif occupancy plot.
}
\description{
This function takes the output from motif density calculations and generates
a plot of relative footprint density across positions relative to a specified motif.
It supports handling multiple samples and grouping them if group information is provided.
}
\examples{
\dontrun{rel_dist_motif_plot(motif_occupancy_file = c("sample1.txt", "sample2.txt"),
                    sample_name = c("sample1", "sample2"),
                    group_name = c("group1", "group1"),
                    motif = "PKP",
                    site = "P")}


}
