% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribo_QC_plot.R
\name{rel_to_start_stop}
\alias{rel_to_start_stop}
\alias{rel_to_start_stop,ribosomeObj-method}
\title{Plot the distribution of reads relative to start or stop codon}
\usage{
rel_to_start_stop(
  object,
  assay = c("raw", "shifted"),
  type = c("relst", "relsp"),
  dist_range = NULL,
  shift = 0,
  geom_col_list = list(),
  facet_wrap_list = list(),
  ...
)

\S4method{rel_to_start_stop}{ribosomeObj}(
  object,
  assay = c("raw", "shifted"),
  type = c("relst", "relsp"),
  dist_range = NULL,
  shift = 0,
  geom_col_list = list(),
  facet_wrap_list = list(),
  ...
)
}
\arguments{
\item{object}{ribosomeObj object.}

\item{assay}{Using raw data or shifted data to plot, c("raw","shifted").}

\item{type}{A character string specifying the type of distribution. Must be
one of "relst" (relative to start codon) or "relsp" (relative to stop codon).}

\item{dist_range}{A numeric vector specifying the minimum and maximum distances
to include in the plot. By default, the entire range(-50-100/-100-50) is included.}

\item{shift}{A numeric value specifying the p site to shift, default 0.}

\item{geom_col_list}{A list of arguments to pass to the
\code{\link[ggplot2]{geom_col}} layer.}

\item{facet_wrap_list}{A list of arguments to pass to the
\code{\link[ggplot2]{facet_wrap}} layer.}

\item{...}{Useless args.}
}
\value{
A ggplot object displaying the distribution of reads relative to the
start or stop codon.
}
\description{
This function takes in a data frame containing quality control data for a set
of samples, and plots the
distribution of reads relative to either the start or stop codon. The user can
specify the type of distribution,as well as a distance range to filter the data.
The function returns a ggplot object.
}
\examples{
\dontrun{
rel_to_start_stop(qc_data = my_qc_data, type = "relst", dist_range = c(-50, 100))
}
}
