% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.class.R
\docType{class}
\name{ribosomeObj-class}
\alias{ribosomeObj-class}
\alias{ribosomeObj}
\title{ribosomeObj Class}
\description{
A S4 class that represents ribosome profiling data.
}
\details{
This class is used for storing and manipulating ribosome profiling data.
The class contains several slots that store raw, shifted, and normalized data, as well as
annotation information related to coding sequences and their corresponding amino acid sequences.
}
\section{Slots}{

\describe{
\item{\code{raw.counts}}{A data.frame containing raw ribosome profiling counts.}

\item{\code{shifted.data}}{A data.frame containing ribosome profiling data after E/P/A-site shifting.}

\item{\code{normalized.data}}{A data.frame containing normalized ribosome profiling data.}

\item{\code{enriched.data}}{A data.frame containing enrichment data for seRP data.}

\item{\code{longest.annotation}}{A data.frame containing the longest annotated sequences.}

\item{\code{shifted}}{A character vector indicating whether the data has been E/P/A-site shifted.}

\item{\code{normalized}}{A character vector indicating whether the data has been normalized.}

\item{\code{genome.file}}{A character vector indicating the path of genome fasta file.}

\item{\code{longest.anno.file}}{A character vector indicating the path of longest transcript annotation file.}

\item{\code{CDS.sequence}}{A character vector containing the coding DNA sequences (CDS).}

\item{\code{amino.acid.sequence}}{A character vector containing the amino acid sequences derived from the CDS.}
}}

\examples{
\dontrun{
# Create an empty ribosomeObj object
ribo <- new("ribosomeObj",
            raw.counts = data.frame(),
            shifted.data = data.frame(),
            normalized.data = data.frame(),
            longest.annotation = data.frame(),
            shifted = "",
            normalized = "",
            genome.file = "",
            longest.anno.file = "",
            CDS.sequence = "",
            amino.acid.sequence = "")
}

}
