% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{seq_frame_plot}
\alias{seq_frame_plot}
\title{Plot Sequence Frame}
\usage{
seq_frame_plot(
  cds_seq = NULL,
  seq_mark = c("both", "start", "stop"),
  trans_id = NULL
)
}
\arguments{
\item{cds_seq}{A character string or Biostrings object representing the coding DNA sequence (CDS).}

\item{seq_mark}{A character string specifying the type of codon to annotate:
\itemize{
  \item \code{"both"} for both start and stop codons.
  \item \code{"start"} for start codons only.
  \item \code{"stop"} for stop codons only.
}
Default is \code{"both"}.}

\item{trans_id}{A character string or integer specifying the transcript ID for which the CDS sequence should be plotted.}
}
\value{
A list containing:
  \itemize{
    \item \code{plot}: A ggplot2 object representing the sequence frame plot.
    \item \code{df_pos}: A data frame containing the positions of start and stop codons in the sequence.
  }
}
\description{
This function plots the sequence frame of a given CDS (Coding DNA Sequence) with
annotations for start and stop codons. The function can be used to visualize the
positions of these codons relative to the reading frame of the transcript.
}
