% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_gene_plot.R
\name{single_gene_plot}
\alias{single_gene_plot}
\alias{single_gene_plot,ribosomeObj-method}
\title{Single Gene Plot}
\usage{
single_gene_plot(
  object,
  select_gene = NULL,
  merge_rep = FALSE,
  mode = c("nt", "codon"),
  aes_col = c("grey", "red"),
  rna_track_df = NULL,
  scale_RNA_factor = 1,
  structure_position = c("top", "bottom"),
  orf_col = "grey65",
  utr_col = "grey",
  utr_width = 1,
  cds_width = 8,
  cds_region_scale_x = 0.1,
  cds_region_anno = NULL,
  cds_anno_col = "#996600",
  anno_region_width = 5,
  geom_col_params = list(),
  ...
)

\S4method{single_gene_plot}{ribosomeObj}(
  object,
  select_gene = NULL,
  merge_rep = FALSE,
  mode = c("nt", "codon"),
  aes_col = c("grey", "red"),
  rna_track_df = NULL,
  scale_RNA_factor = 1,
  structure_position = c("top", "bottom"),
  orf_col = "grey65",
  utr_col = "grey",
  utr_width = 1,
  cds_width = 8,
  cds_region_scale_x = 0.1,
  cds_region_anno = NULL,
  cds_anno_col = "#996600",
  anno_region_width = 5,
  geom_col_params = list(),
  ...
)
}
\arguments{
\item{object}{An object of class `ribosomeObj` containing normalized ribosome
profiling data and gene annotations.}

\item{select_gene}{A character string specifying the gene of interest to plot.
Must correspond to a gene name in the provided object.}

\item{merge_rep}{Logical. If `TRUE`, replicates will be merged by taking the mean RPM. Default is `FALSE`.}

\item{mode}{Character. Determines the unit of x-axis as "nt" for nucleotides
or "codon" for codons/amino acids. Default is `"nt"`.}

\item{aes_col}{Fill color for RNA and Ribo track, c("grey","red").}

\item{rna_track_df}{A data frame contains RNA coverage data which from load_rna_coverage function.}

\item{scale_RNA_factor}{Relative scale density size factor for RNA for better visualiztion. Default is 1.}

\item{structure_position}{Character. Defines the position of transcript
structure ("top" or "bottom"). Default is `"top"`.}

\item{orf_col}{Color code for the coding sequence (CDS) region in the plot. Default is `"grey65"`.}

\item{utr_col}{Color code for the untranslated regions (UTRs) in the plot. Default is `"grey"`.}

\item{utr_width}{Numeric. The line width for UTRs in the plot. Default is `1`.}

\item{cds_width}{Numeric. The line width for CDS in the plot. Default is `8`.}

\item{cds_region_scale_x}{Numeric. Scales the x-axis for the CDS region. Default is `0.1`.}

\item{cds_region_anno}{Data frame. An optional annotation for the CDS region to be plotted.}

\item{cds_anno_col}{Color code for CDS annotations. Default is `"#996600"`.}

\item{anno_region_width}{Numeric. The width for the annotated regions. Default is `5`.}

\item{geom_col_params}{List. Additional parameters to be passed to `geom_col`
for customizing the bar plot aesthetics.}

\item{...}{Other arguments passed to internal plotting functions.}
}
\value{
A ggplot object displaying the ribosome footprint density across the selected gene transcript.
}
\description{
This method generates a plot showing ribosome positioning along a transcript for a specific gene.
}
\examples{
\dontrun{# Assume 'ribosome_data' is a preloaded object of class 'ribosomeObj'
single_gene_plot(ribosome_data, select_gene = "GCN4", mode = "codon", merge_rep = TRUE)}

}
