% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboseqTools_implementation.R
\name{single_gene_profile}
\alias{single_gene_profile}
\title{Single-gene Ribosome Profiling Plot}
\usage{
single_gene_profile(
  data = NULL,
  transform = "log2",
  aes_color = "sample",
  y_threshold = 1,
  shadow_alpha_range = c(0.2, 1),
  facet_grid_params = list(),
  add_cds_structure = TRUE,
  cds_col = "grey",
  cds_anno_col = "#996600",
  cds_region_width = 5,
  cds_region_scale_x = NULL
)
}
\arguments{
\item{data}{A dataframe containing ribosome profiling data. The attribute "bin"
should indicate the bin type (`"byaa"` or `"bynuc"`), and the attribute "type"
should determine if the data represents `"enrichment"` or ribosome density.}

\item{transform}{Character string determining the scale transformation for the y-axis.
Default is `"log2"`.}

\item{aes_color}{Aesthetics mapping for color, usually representing a grouping variable
like `"sample"` or `"rep"`. Default is `"sample"`.}

\item{y_threshold}{Numeric threshold for drawing a horizontal dashed line when the
plot type is enrichment. Default is `1`.}

\item{shadow_alpha_range}{A numeric vector of length 2 controlling the range of the
alpha transparency for shadow regions. Default is `c(0.2,1)`.}

\item{facet_grid_params}{A list of additional parameters passed to `facet_grid`.
Default is an empty list.}

\item{add_cds_structure}{Logical, whether to add a coding sequence (CDS) annotation
structure to the plot. Default is `TRUE`.}

\item{cds_col}{Color of the CDS annotation line. Default is `"grey"`.}

\item{cds_anno_col}{Color of the CDS annotation labels. Default is `"#996600"`.}

\item{cds_region_width}{Numeric, the width of the CDS annotation line. Default is `5`.}

\item{cds_region_scale_x}{Numeric, scaling factor for the x-axis of the CDS region
annotation. Default is `NULL` (no scaling).}
}
\value{
A ggplot object with ribosome profiling visualizations, optionally including
  structure annotations like the CDS.
}
\description{
This function creates a visualization for ribosome profiling of a single gene,
supporting clustering, functional annotation, and publication-ready plots.
The input data can be transformed and structured into a faceted plot,
with optional inclusion of coding sequence (CDS) annotations.
}
