% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_generic_methods.R
\name{subdata}
\alias{subdata}
\alias{subdata,ribosomeObj-method}
\title{Subset data from ribosomeObj object.}
\usage{
subdata(object, ...)

\S4method{subdata}{ribosomeObj}(object, ...)
}
\arguments{
\item{object}{ribosomeObj object.}

\item{...}{Useless args.}
}
\value{
ribosomeObj object.
}
\description{
Subset data from ribosomeObj object.

method for subdata
}
