% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlottingFunctions.R
\name{track_plot}
\alias{track_plot}
\title{Plot ribosome density and RNA coverage along transcript}
\usage{
track_plot(
  signal_data = NULL,
  gene_anno = NULL,
  plot_type = c("translatome", "interactome"),
  add_gene_struc = TRUE,
  show_cds_region_only = FALSE,
  structure_label_y = -3,
  structure_label_size = 2.5,
  structure_col = NULL,
  background_col = "white",
  range_pos = c(0.8, 0.85),
  signal_range_digits = 0,
  reverse_rna = TRUE,
  rna_signal_scale = 1,
  signal_col = NULL,
  line_col = NULL,
  panel_size = NULL,
  remove_all_panel_border = FALSE,
  remove_trans_panel_border = TRUE,
  gene_order = NULL,
  sample_order = NULL,
  fixed_col_range = TRUE,
  show_ribo_only = FALSE,
  show_x_ticks = FALSE,
  show_y_ticks = FALSE,
  sample_group_info = NULL,
  geom_col_params = list(),
  geom_line_params = list(),
  geom_hline_params = list(),
  geom_ribbon_params = list()
)
}
\arguments{
\item{signal_data}{A data frame with columns "transpos", "density", "type"
(either "ribo" or "rna"), "gene_name", "sample".}

\item{gene_anno}{Longest transcript gene annotation file.}

\item{plot_type}{the plot type for track, "translatome"(default) or "interactome".}

\item{add_gene_struc}{Whether add gene structure plot. Default is TRUE.}

\item{show_cds_region_only}{Whether retain CDS region only. Default is FALSE.}

\item{structure_label_y}{The y position of gene structure, defalut -3.}

\item{structure_label_size}{The text size of gene structure labels, defalut 2.5.}

\item{structure_col}{Color of the gene structure rectangles.}

\item{background_col}{Color of the plot background. Default is "grey90".}

\item{range_pos}{Position of the range labels. Default is c(0.85,0.85).}

\item{signal_range_digits}{The signal range digits, you can set it bigger when
signal range is small than 1, defalut 0.}

\item{reverse_rna}{Whether reverse RNA track. Default is TRUE.}

\item{rna_signal_scale}{The scale value of rna coverage track. Default is 1.
It is used for better visualization when the difference of ribo density value
and rna coverage value is quite huge.}

\item{signal_col}{A named vector of colors for ribosome and RNA signals.
Default is NULL, which means using a default color scheme.}

\item{line_col}{A named vector of colors for enrichment line signals.
Default is NULL, which means using a default color scheme.}

\item{panel_size}{A numeric vector of length 2 specifying the width and height
of each panel. Default is NULL, which means using a default panel size.}

\item{remove_all_panel_border}{Logical indicating whether to remove all panel
borders. Default is FALSE.}

\item{remove_trans_panel_border}{Logical indicating whether to remove the
border of the transcript panel only. Default is TRUE.}

\item{gene_order}{An optional character vector specifying the order of genes
in the plot. By default, genes are ordered by their appearance in the input data.}

\item{sample_order}{An optional character vector specifying the order of
samples in the plot. By default, samples are ordered by their appearance in
the input data.}

\item{fixed_col_range}{Logical indicating whether to use a fixed color range
for all panels. Default is TRUE.}

\item{show_ribo_only}{Show only ribo density track. Default is FALSE.}

\item{show_x_ticks}{show_x_ticks Whether show x aixs ticks. Default is FALSE.}

\item{show_y_ticks}{Whether show y aixs ticks. Default is FALSE.}

\item{sample_group_info}{The groups for samples, giving a named list with
samples, default NULL.}

\item{geom_col_params}{The parameters passed by ggplot2::geom_col function for
"translatome" plot type to overwrite params.}

\item{geom_line_params}{The parameters passed by ggplot2::geom_line function for
"interactome" plot type to overwrite params.}

\item{geom_hline_params}{The parameters passed by ggplot2::geom_hline function for
"interactome" plot type to overwrite params.}

\item{geom_ribbon_params}{The parameters passed by ggplot2::geom_ribbon function for
"interactome" plot type to overwrite params.}
}
\description{
This function plots ribosome density and RNA coverage along the transcript,
with gene structures overlaid. The input data should be a data frame with
columns "transpos", "density", "type" (either "ribo" or "rna"), "gene_name",
"sample". Gene structures are loaded from a file named "longest_info.txt".
You can customize various aspects of the plot, such as color scheme, panel
size, and panel borders.
}
