% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RiboseqTools_implementation.R
\name{transform_to_sparse_matrix}
\alias{transform_to_sparse_matrix}
\title{Transform ribosomeObj Object of shifted.data into sparsed matrix}
\usage{
transform_to_sparse_matrix(object = NULL, sample_name = NULL)
}
\arguments{
\item{object}{ribosomeObj Object.}

\item{sample_name}{sample name.}
}
\value{
A sparsed matrix containing gene read counts on cds region.
}
\description{
Transform ribosomeObj Object of shifted.data into sparsed matrix
}
