% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripeptide_motif_score.R
\name{triAmino_motif_plot}
\alias{triAmino_motif_plot}
\title{Plot Tri-amino Acid Motifs}
\usage{
triAmino_motif_plot(
  occupancy_file = NULL,
  sample_name = NULL,
  group_name = NULL,
  top_motif = NULL
)
}
\arguments{
\item{occupancy_file}{Vector of strings; paths to the occupancy files containing
motifs and their scores.}

\item{sample_name}{Vector of strings; names of the samples corresponding to each
file. Used for labeling in the absence of a group name.}

\item{group_name}{Optional; vector of strings providing group names for each sample.
If provided, scores are averaged across groups.}

\item{top_motif}{Integer; number of top motifs to display based on their ratio.}
}
\value{
A ggplot object displaying the sequence logo of top motifs based on the
        calculated ratio.
}
\description{
This function takes motif occupancy data and generates a plot using the ggseqlogo
package to visualize the frequency or probability of amino acid motifs. It handles
multiple samples and can average scores across specified groups.
}
\examples{
\dontrun{triAmino_motif_plot(occupancy_file = c("path/to/file1.txt", "path/to/file2.txt"),
                    sample_name = c("sample1", "sample2"),
                    group_name = c("group1", "group1"),
                    top_motif = 5)}

}
