% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripeptide_motif_score.R
\name{triAmino_scater_plot}
\alias{triAmino_scater_plot}
\title{Create a Scatter Plot of Tripeptide Occupancy Scores}
\usage{
triAmino_scater_plot(
  occupancy_file = NULL,
  sample_name = NULL,
  group_name = NULL,
  mark_motif = NULL,
  mark_color = "orange"
)
}
\arguments{
\item{occupancy_file}{A vector of file paths to tripeptide occupancy score files.}

\item{sample_name}{A vector of sample names corresponding to each occupancy file.}

\item{group_name}{An optional vector of group names for each sample.}

\item{mark_motif}{A vector of motifs to highlight in the plot.}

\item{mark_color}{The color to use for highlighted motifs. Default is "orange".}
}
\value{
A ggplot object representing the scatter plot of tripeptide occupancy scores.
}
\description{
This function generates a scatter plot comparing tripeptide occupancy scores between two samples or groups.
}
\details{
This function performs the following steps:
  1. Reads and combines tripeptide occupancy data from multiple files.
  2. Calculates mean scores if group names are provided.
  3. Creates a scatter plot comparing scores between two samples or groups.
  4. Highlights specified motifs with labels and different colors.
}
\examples{
\dontrun{triAmino_scater_plot(
  occupancy_file = c("sample1_occupancy.txt", "sample2_occupancy.txt"),
  sampe_name = c("Sample1", "Sample2"),
  mark_motif = c("PPP", "GGG"),
  mark_color = "red"
)}

}
