% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WholeMetagenePlot.R
\name{whole_metagene_plot}
\alias{whole_metagene_plot}
\title{Create a Whole Metagene Plot}
\usage{
whole_metagene_plot(
  gene_file = NULL,
  qc_data = NULL,
  scale_region = FALSE,
  bins = NULL,
  geom_line_list = list(),
  facet_wrap_list = list()
)
}
\arguments{
\item{gene_file}{A character string specifying the path to the gene annotation file
which is from pre_longest_trans_info function. Defaults to `NULL`.}

\item{qc_data}{A data frame containing QC data with transcript information.
which is from load_qc_data function. Defaults to `NULL`.}

\item{scale_region}{A logical value indicating whether to scale the UTR regions
relative to the CDS. Defaults to `FALSE`.}

\item{bins}{Two length of numerical vector for region scale bins. Default NULL.}

\item{geom_line_list}{A list of arguments to pass to the `geom_line` function
in `ggplot2`. Defaults to an empty list.}

\item{facet_wrap_list}{A list of arguments to pass to the `facet_wrap` function
in `ggplot2`. Defaults to an empty list.}
}
\value{
A ggplot object representing the whole metagene plot.
}
\description{
This function generates a whole metagene plot based on gene annotation and
QC data. It calculates the relative distance of reads from the transcription
start site and plots the density of read footprints across different regions
of transcripts (5'UTR, CDS, and 3'UTR).
}
