% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_reads_offset.R
\name{adjust_offset}
\alias{adjust_offset}
\alias{adjust_offset,ribosomeObj-method}
\title{Adjust Offset for RPF}
\usage{
adjust_offset(object, offset_df = NULL, shift = 0, ...)

\S4method{adjust_offset}{ribosomeObj}(object, offset_df = NULL, shift = 0, ...)
}
\arguments{
\item{object}{ribosomeObj object}

\item{offset_df}{A data frame containing the read lengths and offsets.}

\item{shift}{A shift for E/P/A site of reads position adjustment. Default 0.}

\item{...}{useless args}
}
\value{
A data frame with adjusted relative start positions.
}
\description{
This function adjusts the offsets for quality control data based on the read lengths and associated offsets.
It separates the read lengths and offsets, performs necessary data type conversions, and then adjusts
the relative start positions by the absolute values of the offsets.
}
