% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_occupancy.R
\name{codon_occupancy}
\alias{codon_occupancy}
\title{Calculate Codon Occupancy}
\usage{
codon_occupancy(
  qc_file = NULL,
  longest_trans_file = NULL,
  min_counts = 32,
  cds_fasta_file = NULL,
  upstream_codon_exclude = 5,
  downstream_codon_exclude = 0
)
}
\arguments{
\item{qc_file}{A data frame containing quality control data, including columns for sample, transcript ID,
transcript position, counts, and normalized expression.}

\item{longest_trans_file}{A string specifying the path to the file containing the longest transcript information.
This file should contain columns for transcript ID, gene name, gene ID, chromosome, strand,
CDS range, exon range, 5' UTR length, CDS length, and 3' UTR length.}

\item{min_counts}{An integer specifying the minimum number of counts required for a transcript to be included
in the analysis. Default is 32.}

\item{cds_fasta_file}{A string specifying the path to the FASTA file containing CDS sequences.}

\item{upstream_codon_exclude}{An integer specifying the number of codons to exclude from the start of the coding sequence.
Default is 5.}

\item{downstream_codon_exclude}{An integer specifying the number of codons to exclude from the end of the coding sequence.
Default is 0.}
}
\value{
This function does not return a value, but creates output files in a 'codon_occupancy' directory:
  - A text file for each sample containing codon position expression data.
  - A text file for each sample containing codon occupancy data.
}
\description{
This function calculates codon occupancy based on ribosome profiling data and transcript information.
}
\details{
This function performs the following steps:
  1. Filters reads based on minimum count threshold.
  2. Calculates codon density for each transcript.
  3. Outputs codon position expression data.
  4. Calls a Python script to calculate codon occupancy.
}
