% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_track_plot.R
\name{codon_track_plot}
\alias{codon_track_plot}
\title{Generate Codon Track Plot}
\usage{
codon_track_plot(
  codon_exp_file = NULL,
  longest_trans_file = NULL,
  sample_name = NULL,
  group_name = NULL,
  select_gene = NULL,
  geom_col_list = list()
)
}
\arguments{
\item{codon_exp_file}{A vector of file paths to codon expression data files.}

\item{longest_trans_file}{A string specifying the path to the file containing the longest transcript information.}

\item{sample_name}{A vector of sample names corresponding to each codon expression file.}

\item{group_name}{An optional vector of group names for each sample.}

\item{select_gene}{A vector of gene names to be plotted.}

\item{geom_col_list}{A list of additional parameters to be passed to geom_col().}
}
\value{
A ggplot object representing the combined track plot for all selected genes.
}
\description{
This function creates a track plot of codon-level ribosome footprint density for selected genes.
}
\details{
This function performs the following steps:
  1. Loads codon expression data from multiple files.
  2. Processes gene annotation data.
  3. Filters and combines data for selected genes.
  4. Creates individual track plots for each gene.
  5. Combines all plots into a single ggplot object.

The resulting plot shows ribosome footprint density along transcript positions for each selected gene,
with separate tracks for each sample or group.
}
\examples{
\dontrun{codon_track_plot(
  codon_exp_file = c("sample1_codon_exp.txt", "sample2_codon_exp.txt"),
  longest_trans_file = "longest_transcripts.txt",
  sample_name = c("Sample1", "Sample2"),
  select_gene = c("GENE1", "GENE2"),
  geom_col_list = list(width = 0.8, position = "dodge")
)}

}
