% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/footprint_heatmap.R
\name{footprint_heatmap}
\alias{footprint_heatmap}
\title{Create a footprint heatmap}
\usage{
footprint_heatmap(
  qc_data = NULL,
  type = c("relst", "relsp"),
  rel_distance = NULL,
  read_length = c(20, 35),
  tile_border = "grey",
  bar_col = "grey",
  bar_width = 0.75,
  scale_x = 0.1,
  scale_y = 0.1,
  xseq = 3,
  xline = FALSE,
  xline_col = "black",
  log = FALSE
)
}
\arguments{
\item{qc_data}{A data frame containing the quality control data.}

\item{type}{A character string specifying the type of distance to use.
It can be either 'relst' or 'relsp'. Default is 'relst'.}

\item{rel_distance}{A numeric vector specifying the relative distance
from the 5' end to the start or stop codon. Default is NULL.}

\item{read_length}{A numeric vector specifying the read length range.
Default is c(20,35).}

\item{tile_border}{A character string specifying the color of the tile
border. Default is "grey".}

\item{bar_col}{A character string specifying the color of the bar.
Default is "grey".}

\item{bar_width}{A numeric value specifying the width of the bar.}

\item{scale_x}{X panel size. Default is 0.1.}

\item{scale_y}{Y panel size. Default is 0.1.}

\item{xseq}{The x axis interval. Default is 3.}

\item{xline}{Whether show x side plot with line plot. Default is FASLE.}

\item{xline_col}{The color of x side line when xline = TRUE. Default is "black".
Default is 0.75.}

\item{log}{Whether do log transformation for counts. Default is FALSE.}
}
\value{
A ggplot2 object representing the footprint heatmap.
}
\description{
This function creates a footprint heatmap for ribo-seq data. The heatmap
visualizes the count of reads at different distances from the start
or stop codon relative to the 5' end of the transcript. The type of
distance can be specified as 'relst' for the distance from the start
codon or 'relsp' for the distance from the stop codon.
}
