% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nomalized_counts.R
\name{get_nomalized_counts}
\alias{get_nomalized_counts}
\title{Get Normalized Counts for Transcript Positions}
\usage{
get_nomalized_counts(
  longest_trans_file = NULL,
  qc_df = NULL,
  exclude_upstreamCodon = 30,
  exclude_downstreamCodon = 0
)
}
\arguments{
\item{longest_trans_file}{A string specifying the path to the file containing the longest transcript information.
This file should contain columns for transcript ID, gene name, gene ID, chromosome, strand,
CDS range, exon range, 5' UTR length, CDS length, and 3' UTR length.}

\item{qc_df}{A data frame containing quality control data, including columns for sample, transcript ID,
transcript position, and counts.}

\item{exclude_upstreamCodon}{An integer specifying the number of codons to exclude from the start of the coding sequence.
Default is 30.}

\item{exclude_downstreamCodon}{An integer specifying the number of codons to exclude from the end of the coding sequence.
Default is 0.}
}
\value{
A data frame with normalized counts for each position in each transcript across all samples.
        The returned data frame includes all original columns from qc_df plus a new column 'norm_exp'
        representing the normalized expression.
}
\description{
This function normalizes the counts at individual positions to the average counts of the transcript,
with options to exclude a specified number of codons from the start and end of the coding sequence.
}
\details{
The function performs the following steps:
  1. Reads and processes the gene annotation file.
  2. Calculates total reads and average expression for each transcript, excluding specified regions.
  3. Normalizes the counts at each position by dividing by the average expression of the transcript.
}
