% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pausing_site.R
\name{get_pausing_site}
\alias{get_pausing_site}
\title{Detect Pausing Sites}
\usage{
get_pausing_site(
  longest_trans_file = NULL,
  normed_file = NULL,
  min_counts = 64,
  window = 500
)
}
\arguments{
\item{longest_trans_file}{A character string specifying the file path to the
longest transcripts file. Default is \code{NULL}.}

\item{normed_file}{A data frame containing normalized expression data. Default is \code{NULL}.}

\item{min_counts}{An integer specifying the minimum counts threshold for
considering a site as a pausing site. Default is \code{64}.}

\item{window}{An integer specifying the window size for detecting pausing
sites. Default is \code{500}.}
}
\value{
This function does not return a value. It writes output files to the
"pausing_site_data" directory.
}
\description{
This function processes normalized ribosome profiling data to detect pausing sites.
It writes intermediate normed data files and uses a Python script to perform
the actual detection of pausing sites. This method can be refered to pausepred
[PausePred and Rfeet: webtools for inferring ribosome pauses and visualizing footprint density from ribosome profiling data](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6140459/)
}
\examples{
\dontrun{
# Example usage:
get_pausing_site(longest_trans_file = "longest_transcripts.txt",
                normed_file = normed_data_frame,
                min_counts = 64,
                window = 500)
}
}
