% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_occupancy.R
\name{get_track_df}
\alias{get_track_df}
\title{Generate Track Data Frame for Selected Genes}
\usage{
get_track_df(longest_trans_file = NULL, select_gene = NULL, normed_file = NULL)
}
\arguments{
\item{longest_trans_file}{A string specifying the file path to the longest transcript data, which includes gene annotations. Expected to be a delimited file without headers. Default is NULL.}

\item{select_gene}{A character vector of gene names to filter the gene annotation data. If NULL, no filtering is applied. Default is NULL.}

\item{normed_file}{A data frame or a string specifying the file path to the normalized gene expression data, which should include `trans_id` and `norm_exp` columns among others. Default is NULL.}
}
\value{
A data frame containing the filtered and transformed data suitable for tracking gene expression. The returned data frame includes columns for sample, gene name, transcript ID, transcript position (transpos), normalized expression (density), and a fixed column (type) set to 'ribo'.
}
\description{
This function extracts and prepares track data for selected genes from gene annotation and normalized expression files.
It filters gene annotation data for selected genes, merges it with normalized expression data, and formats it for further analysis or visualization.
}
\examples{
\dontrun{# Assuming `longest_trans_file` and `normed_file` are already defined:
track_df <- get_track_df(
  longest_trans_file = "path/to/longest_transcripts.tsv",
  select_gene = c("Gene1", "Gene2"),
  normed_file = expression_data_frame
)}

}
