% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tripeptide_motif_score.R
\name{peptide_motif_score2}
\alias{peptide_motif_score2}
\title{Peptide Motif Scoring Function}
\usage{
peptide_motif_score2(
  amino_file = NULL,
  normed_exp_file = NULL,
  longest_trans_file = NULL,
  norm_type = c("normed_count", "average_count"),
  window = 50,
  occurrence_threshold = 100
)
}
\arguments{
\item{amino_file}{A character string specifying the path to the amino acid file.}

\item{normed_exp_file}{A character string specifying the path to the normalized expression file.}

\item{longest_trans_file}{A character string specifying the path to the longest transcript file.}

\item{norm_type}{A character string specifying the normalization type. It can be one of
"normed_count" (normalized count) or "average_count" (average count).
Default is "normed_count".}

\item{window}{An integer specifying the window size for motif scanning.
Default is 50.}

\item{occurrence_threshold}{An integer specifying the threshold for motif occurrences.
Default is 100.}
}
\value{
The function creates output files with motif scores and prints a message for each processed sample.
}
\description{
This function calculates the motif scores for peptides based on provided files.
}
\examples{
\dontrun{# Example usage:
peptide_motif_score2(amino_file = "path/to/amino_file.txt",
                     normed_exp_file = "path/to/normed_exp_file.txt",
                     longest_trans_file = "path/to/longest_trans_file.txt")}

}
